move logger to debug/ dir to not be counted in codecov
